refactor starknet provider
