+++ client/src/App.tsx
@@ -1,9 +1,8 @@
-import { goerli, mainnet } from '@starknet-react/chains'
-import { argent, braavos, publicProvider, StarknetConfig, useInjectedConnectors } from '@starknet-react/core'
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
 import Layout from 'src/components/Layout'
 
 import SwapPage from './pages/Swap'
+import StarknetProvider from './providers/StarknetProvider'
 
 const router = createBrowserRouter([
   {
@@ -17,17 +16,9 @@ const router = createBrowserRouter([
 ])
 
 export default function App() {
-  const chains = [goerli, mainnet]
-  const provider = publicProvider()
-  const { connectors } = useInjectedConnectors({
-    recommended: [argent(), braavos()],
-    includeRecommended: 'onlyIfNoConnectors',
-    order: 'random',
-  })
-
   return (
-    <StarknetConfig chains={chains} provider={provider} connectors={connectors}>
+    <StarknetProvider>
       <RouterProvider router={router} />
-    </StarknetConfig>
+    </StarknetProvider>
   )
 }
